Further generalize app config functionality
