+++ api/config/__init__.py
@@ -1,1 +1,2 @@
 from api.config.ConfigReader import ConfigReader
+from api.config.GenericAppConfig import GenericAppConfig
